Refactor helpers into plugins/composables
Reorganized frontend shared utilities by moving `i18n` from `helpers` to `plugins` and `notification` from `helpers` to `composables`. Added new barrel `index.js` exports for both folders and updated `main.js` imports to match the new structure.
