+++ electrostoreFRONT/src/plugins/index.js
@@ -1,0 +1,1 @@
+export * from "./i18n";
